Support `Enum._missing_` hooks
This adds support for `Enum._missing_` hooks. This method is part of
the standard library's `enum` implementation, and is called when a value
isn't one defined directly on the `Enum` type. The method should return
a valid enum member, or `None` if the value is invalid.

This may be used to support case-insensitive enums, or to coerce unknown
values into a single "UNKNOWN" member.