feat: add `repr` kwarg to `field` function and `Field` C struct
Extended the field Python function to accept a new keyword argument `repr` which controls the inclusion of the field in the generated `__repr__` string for the struct.

Added a new attribute `repr` to the `Field` C struct to store the value of the `repr` keyword argument.

Updated the `Field` C struct docstring to include details about the new `repr` kwarg.