Use positional-only args in certain function sigs
Previously the type signatures noted for the json and msgpack modules
weren't strictly accurate. The implementation only accepted positional
args for certain parameters, but the signatures defined them as
positional-or-keyword parameters. This PR rectifies that.

Note that this is _not_ a change in behavior, this is only a change in
the signatures that tools like `mypy` or `pyright` would use to check
your code.