Add `raw_decode` method to JSON and MsgPack decoders
Adds a new `raw_decode` method on the JSON and MsgPack `Decoder` classes
that allow decoding objects with trailing data after them.  This mirrors
the interface of the `raw_decode` method in the Python standard
library's `json.JSONDecoder` class.

This can be used to, for example, parse several concatenated messages,
enabling similar functionality to that available through
`msgpack.Unpacker` in the `msgpack` package.